+++ src/contracts/deposits-registry/deposits-registry.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { SecurityModule } from 'contracts/security';
 import { DepositsRegistryStoreModule } from './store';
 import { DepositRegistryService } from './deposits-registry.service';
-import { DEPOSIT_CACHE_DEFAULT } from './deposit-registry.constants';
+import { DEPOSIT_CACHE_DEFAULT } from './deposits-registry.constants';
 import { DepositsRegistryFetcherModule } from './fetcher';
 import { DepositRegistrySanityCheckerModule } from './sanity-checker';
