+++ src/contracts/deposits-registry/index.ts
@@ -1,0 +1,3 @@
+export * from './deposits-registry.module';
+export * from './deposits-registry.service';
+export * from './interfaces';
